DEV-1855: use psycopg2 instead of psycopg2-binary (#60)
Switch from psycopg2-binary to psycopg2. The major reason is for consistency across the rest of the projects. psqlgraph requires psycopg2 and psqlgraph is widely used. Whenever psqlgraph and gdc-ng-models are used in the same project both psycopg2 and psycopg2-binary will be included in the dependencies which is not great.